Refactoring - lesson 5 done
